requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
djoser==2.3.3

channels==4.1.0
uvicorn==0.29.0


psycopg2-binary==2.9.9

Pillow==12.0.0
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-countries==7.5.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-resized==1.0.2
django-templated-mail==1.1.1

whitenoise==6.6.0
gunicorn==21.2.0

requests==2.31.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
